# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/user": {
            "post": {
                "description": "Request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Request"
                ],
                "summary": "Request",
                "operationId": "RequestBody",
                "parameters": [
                    {
                        "description": "RequestBody",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "GetBookBody",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAll"
                        }
                    },
                    "400": {
                        "description": "Invalid Argumant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.RequestBody": {
            "type": "object",
            "properties": {
                "detailNumPlaces": {
                    "type": "integer"
                },
                "direction": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Direction"
                    }
                },
                "showWithoutPlaces": {
                    "type": "integer"
                },
                "stationFrom": {
                    "type": "string"
                },
                "stationTo": {
                    "type": "string"
                },
                "time": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "wagon": {
                    "$ref": "#/definitions/models.WagonType"
                }
            }
        },
        "models.ResponseAll": {
            "type": "object",
            "properties": {
                "discount": {
                    "type": "array",
                    "items": {}
                },
                "express": {
                    "type": "object",
                    "properties": {
                        "code": {},
                        "content": {
                            "type": "string"
                        },
                        "direction": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "notAllTrains": {},
                                    "passRoute": {
                                        "type": "object",
                                        "properties": {
                                            "codeFrom": {
                                                "type": "string"
                                            },
                                            "codeTo": {
                                                "type": "string"
                                            },
                                            "from": {
                                                "type": "string"
                                            },
                                            "to": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "trains": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "date": {
                                                    "type": "string"
                                                },
                                                "train": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "arrival": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "date": {
                                                                        "type": "string"
                                                                    },
                                                                    "localDate": {
                                                                        "type": "string"
                                                                    },
                                                                    "localTime": {
                                                                        "type": "string"
                                                                    },
                                                                    "stop": {
                                                                        "type": "string"
                                                                    },
                                                                    "time": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            },
                                                            "brand": {
                                                                "type": "string"
                                                            },
                                                            "bus": {},
                                                            "comments": {},
                                                            "departure": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "localDate": {
                                                                        "type": "string"
                                                                    },
                                                                    "localTime": {
                                                                        "type": "string"
                                                                    },
                                                                    "stop": {},
                                                                    "time": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            },
                                                            "departureTrain": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "date": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            },
                                                            "elRegPossible": {
                                                                "type": "string"
                                                            },
                                                            "firmName": {},
                                                            "length": {
                                                                "type": "string"
                                                            },
                                                            "number": {
                                                                "type": "string"
                                                            },
                                                            "number2": {
                                                                "type": "string"
                                                            },
                                                            "parom": {},
                                                            "places": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "cars": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "type": "object",
                                                                            "properties": {
                                                                                "freeSeats": {
                                                                                    "type": "string"
                                                                                },
                                                                                "indexType": {
                                                                                    "type": "string"
                                                                                },
                                                                                "seats": {
                                                                                    "type": "object",
                                                                                    "properties": {
                                                                                        "freeComp": {},
                                                                                        "seatsDn": {},
                                                                                        "seatsLateralDn": {},
                                                                                        "seatsLateralUp": {},
                                                                                        "seatsUndef": {
                                                                                            "type": "string"
                                                                                        },
                                                                                        "seatsUp": {}
                                                                                    }
                                                                                },
                                                                                "tariffs": {
                                                                                    "type": "object",
                                                                                    "properties": {
                                                                                        "tariff": {
                                                                                            "type": "array",
                                                                                            "items": {
                                                                                                "type": "object",
                                                                                                "properties": {
                                                                                                    "addFood": {},
                                                                                                    "addSigns": {},
                                                                                                    "carrier": {
                                                                                                        "type": "object",
                                                                                                        "properties": {
                                                                                                            "name": {
                                                                                                                "type": "string"
                                                                                                            }
                                                                                                        }
                                                                                                    },
                                                                                                    "classService": {
                                                                                                        "type": "object",
                                                                                                        "properties": {
                                                                                                            "content": {
                                                                                                                "type": "string"
                                                                                                            },
                                                                                                            "service": {
                                                                                                                "type": "object"
                                                                                                            },
                                                                                                            "type": {
                                                                                                                "type": "string"
                                                                                                            }
                                                                                                        }
                                                                                                    },
                                                                                                    "classServiceInt": {},
                                                                                                    "comissionFee": {
                                                                                                        "type": "string"
                                                                                                    },
                                                                                                    "designCar": {},
                                                                                                    "elRegPossible": {
                                                                                                        "type": "object",
                                                                                                        "properties": {
                                                                                                            "akp": {},
                                                                                                            "uk": {}
                                                                                                        }
                                                                                                    },
                                                                                                    "modificators": {},
                                                                                                    "owner": {
                                                                                                        "type": "object",
                                                                                                        "properties": {
                                                                                                            "country": {
                                                                                                                "type": "object",
                                                                                                                "properties": {
                                                                                                                    "code": {
                                                                                                                        "type": "string"
                                                                                                                    },
                                                                                                                    "name": {
                                                                                                                        "type": "string"
                                                                                                                    }
                                                                                                                }
                                                                                                            },
                                                                                                            "railway": {
                                                                                                                "type": "object",
                                                                                                                "properties": {
                                                                                                                    "code": {
                                                                                                                        "type": "string"
                                                                                                                    },
                                                                                                                    "name": {
                                                                                                                        "type": "string"
                                                                                                                    }
                                                                                                                }
                                                                                                            },
                                                                                                            "type": {
                                                                                                                "type": "string"
                                                                                                            }
                                                                                                        }
                                                                                                    },
                                                                                                    "passVok": {
                                                                                                        "type": "object",
                                                                                                        "properties": {
                                                                                                            "stationFrom": {
                                                                                                                "type": "object",
                                                                                                                "properties": {
                                                                                                                    "code": {
                                                                                                                        "type": "string"
                                                                                                                    }
                                                                                                                }
                                                                                                            },
                                                                                                            "stationTo": {}
                                                                                                        }
                                                                                                    },
                                                                                                    "payFood": {},
                                                                                                    "saleOnFour": {},
                                                                                                    "saleOnTwo": {},
                                                                                                    "seats": {
                                                                                                        "type": "object",
                                                                                                        "properties": {
                                                                                                            "freeComp": {},
                                                                                                            "seatsDn": {},
                                                                                                            "seatsLateralDn": {},
                                                                                                            "seatsLateralUp": {},
                                                                                                            "seatsUndef": {
                                                                                                                "type": "string"
                                                                                                            },
                                                                                                            "seatsUp": {}
                                                                                                        }
                                                                                                    },
                                                                                                    "seatsSex": {},
                                                                                                    "selBedding": {},
                                                                                                    "selFood": {},
                                                                                                    "sellingInternetForbidden": {},
                                                                                                    "station": {},
                                                                                                    "swim": {},
                                                                                                    "tariff": {
                                                                                                        "type": "string"
                                                                                                    },
                                                                                                    "tariff2": {},
                                                                                                    "tariffService": {},
                                                                                                    "ud": {}
                                                                                                }
                                                                                            }
                                                                                        }
                                                                                    }
                                                                                },
                                                                                "type": {
                                                                                    "type": "string"
                                                                                },
                                                                                "typeShow": {
                                                                                    "type": "string"
                                                                                }
                                                                            }
                                                                        }
                                                                    }
                                                                }
                                                            },
                                                            "route": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "station": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "type": "string"
                                                                        }
                                                                    }
                                                                }
                                                            },
                                                            "timeInWay": {
                                                                "type": "string"
                                                            },
                                                            "type": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                }
                            }
                        },
                        "hasError": {
                            "type": "boolean"
                        },
                        "reqAddress": {
                            "type": "string"
                        },
                        "reqExpressDateTime": {
                            "type": "string"
                        },
                        "reqExpressZK": {
                            "type": "string"
                        },
                        "reqLocalRecv": {
                            "type": "string"
                        },
                        "reqLocalSend": {
                            "type": "string"
                        },
                        "showWithoutPlaces": {},
                        "type": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}